+++ test/VersionModule.t.sol
@@ -32,7 +32,7 @@ contract VersionModuleTest is HelperContract {
                 (
                     OWNER,
                     IERC20(address(tokenPayment)),
-                    ISnapshotState(address(snapshotEngine)),
+                    ISnapshotSource(address(snapshotEngine)),
                     IRuleEngine(ZERO_ADDRESS),
                     TIME_LIMIT_TO_WITHDRAW
                 )